Restructured the acceptance tests folder
