+++ tests/Chronos.Tests.Acceptance/Flows/Health/HealthTests.cs
@@ -3,7 +3,7 @@ using Chronos.Domain.Management.Roles;
 using Chronos.Tests.Acceptance.Infrastructure;
 using FluentAssertions;
 
-namespace Chronos.Tests.Acceptance.Flows;
+namespace Chronos.Tests.Acceptance.Flows.Health;
 
 [TestFixture]
 [Category("Acceptance")]
