+++ tests/Chronos.Tests.Acceptance/Flows/OrganizationManagement/OrganizationLifecycleTests.cs
@@ -6,7 +6,7 @@ using Chronos.MainApi.Management.Contracts;
 using Chronos.Tests.Acceptance.Infrastructure;
 using FluentAssertions;
 
-namespace Chronos.Tests.Acceptance.Flows;
+namespace Chronos.Tests.Acceptance.Flows.OrganizationManagement;
 
 [TestFixture]
 [Category("Acceptance")]
